add something new

+++ spark_news/src/main/scala/com/luochan/caseclass/LogsCaseclass.scala
@@ -1,4 +1,4 @@
 package com.luochan.caseclass
 
 case class UserRating(userId:Int, movieId:Int, rating:Double)
-case class newsLog(userId:Int, newsId:Int, rayting:Float,timeStamp:String)
+case class newsLog(userId:Int, newsId:Int, rayting:Double,timeStamp:String)
